requirements.txt
arabic-reshaper==2.1.3
asgiref==3.2.7
asn1crypto==1.5.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==37.0.0
cssselect2==0.6.0
Django==3.0.5
django-ajax-tables==1.1.1
django-bootstrap-modal-forms==1.5.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-tables2==2.3.1
et-xmlfile==1.0.1
future==0.18.2
gunicorn==20.0.4
html5lib==1.1
idna==3.3
jdcal==1.4.1
lxml==4.8.0
openpyxl==3.0.3
oscrypto==1.3.0
Pillow==9.1.0
psycopg2-binary==2.8.5
pycparser==2.21
pyHanko==0.13.0
pyhanko-certvalidator==0.19.5
PyPDF3==1.0.6
python-bidi==0.4.2
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
reportlab==3.6.9
requests==2.27.1
six==1.16.0
sqlparse==0.3.1
svglib==1.2.1
tinycss2==1.1.1
tqdm==4.64.0
tzdata==2022.1
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.9
webencodings==0.5.1
xhtml2pdf==0.2.7
